Update EmployeeServiceImplTest for testing KafkaProducer
